[refactor] RoomParticipants 도메인 Room 패키지 하위로 이동 (#80)
